requirements.txt
numpy
tensorflow-gpu
wrapt
tensorflow-probability
opencv-python